+++ mlflow/pyfunc/__init__.py
@@ -90,10 +90,11 @@ import sys
 from copy import deepcopy
 
 import mlflow
+import mlflow.pyfunc.model
 from mlflow.tracking.fluent import active_run, log_artifacts
 from mlflow import tracking
 from mlflow.models import Model
-from mlflow.pyfunc.model import save_model, log_model, PythonModel, PythonModelContext,\
+from mlflow.pyfunc.model import, PythonModel, PythonModelContext,\
     DEFAULT_CONDA_ENV
 from mlflow.utils import PYTHON_VERSION, get_major_minor_py_version
 from mlflow.utils.file_utils import TempDir, _copy_file_or_tree
@@ -153,6 +154,108 @@ def _load_model_env(path, run_id=None):
     if run_id is not None:
         path = tracking.utils._get_model_log_dir(path, run_id)
     return _get_flavor_configuration(model_path=path, flavor_name=FLAVOR_NAME).get(ENV, None)
+
+
+def save_model(dst_path, loader_module=None, data_path=None, code_paths=None, conda_env=None,
+               mlflow_model=Model(), artifacts=None, parameters=None, model_class=None):
+    first_argument_set = {
+        "loader_module": loader_module,
+        "data_path": data_path,
+    }
+    second_argument_set = {
+        "artifacts": artifacts,
+        "parameters": parameters,
+        "model_class": model_class,
+    }
+    first_argument_set_specified = any([item is not None for item in first_argument_set.values()])
+    second_argument_set_specified = any([item is not None for item in second_argument_set.values()])
+    if first_argument_set_specified and second_argument_set_specified:
+        raise MlflowException(
+            "The following sets of arguments cannot be specified together: {first_set_keys}"
+            " and {second_set_keys}. All arguments in one set must be `None`. Instead, found the"
+            " following values: {first_set_entries} and {second_set_entries}".format(
+                first_set_keys=first_argument_set.keys(),
+                second_set_keys=second_argument_set.keys(),
+                first_set_entries=first_argument_set,
+                second_set_entries=second_argument_set))
+
+    if loader_module is not None:
+        return _save_model(
+            dst_path=dst_path, loader_module=loader_module, data_path=data_path,
+            code_paths=code_paths, conda_env=conda_env, mlflow_model=mlflow_model)
+    elif model_class is not None:
+        return mlflow.pyfunc.model._save_model(
+            path=dst_path, model_class=model_class, artifacts=artifacts, parameters=parameters,
+            conda_env=conda_env, code_paths=code_paths, mlflow_model=mlflow_model)
+    elif data_path is not None:
+        raise MlflowException(
+            "`data_path` was specified, but the `loader_module` argument was not provided.")
+    elif artifacts is not None or parameters is not None:
+        raise MlflowException("`artifacts` or `parameters` was specified, but the `model_class`"
+                              " argument was not provided.")
+
+
+def _save_model(dst_path, loader_module, data_path=None, code_paths=None, conda_env=None,
+                mlflow_model=Model()):
+    pass
+
+
+def log_model(artifact_path, model_class, artifacts=None, parameters=None, conda_env=None,
+              code_paths=None):
+    """
+    :param path: The run-relative artifact path to which to log the Python model.
+    :param model_class: A ``type`` object referring to a subclass of
+                        :class:`~PythonModel`, or the fully-qualified name of such a subclass.
+                        ``model_class`` defines how the model is loaded and how it performs
+                        inference.
+    :param artifacts: A dictionary containing ``<name, artifact_uri>`` entries. Remote artifact URIs
+                      will be resolved to absolute filesystem paths, producing a dictionary of
+                      ``<name, absolute_path>`` entries. ``model_class`` can reference these
+                      resolved entries as the ``artifacts`` property of the ``context`` attribute.
+                      For example, consider the following ``artifacts`` dictionary::
+
+                        {
+                            "my_file": "s3://my-bucket/path/to/my/file"
+                        }
+
+                      In this case, the ``"my_file"`` artifact will be downloaded from S3. The
+                      ``model_class`` can then refer to ``"my_file"`` as an absolute filesystem path
+                      via ``self.context.artifacts["my_file"]``.
+    :param parameters: A dictionary containing ``<name, python_object>`` entries. ``python_object``
+                       may be any Python object that is serializable with CloudPickle.
+                       ``model_class`` can reference these resolved entries as the ``parameters``
+                       property of the ``context`` attribute. For example, consider the following
+                       ``parameters`` dictionary::
+
+                         {
+                             "my_list": range(10)
+                         }
+
+                       The ``model_class`` can refer to the Python list named ``"my_list"`` as
+                       ``self.context.parameters["my_list"]``.
+    :param conda_env: Either a dictionary representation of a Conda environment or the path to a
+                      Conda environment yaml file. If provided, this decribes the environment
+                      this model should be run in. At minimum, it should specify the dependencies
+                      contained in :data:`mlflow.pyfunc.DEFAULT_CONDA_ENV`. If `None`, the default
+                      :data:`mlflow.pyfunc.DEFAULT_CONDA_ENV` environment will be added to the
+                      model. The following is an *example* dictionary representation of a Conda
+                      environment::
+
+                        {
+                            'name': 'mlflow-env',
+                            'channels': ['defaults'],
+                            'dependencies': [
+                                'python=3.7.0',
+                                'cloudpickle=0.5.8'
+                            ]
+                        }
+    :param code_paths: A list of paths to Python file dependencies that are required by
+                       instances of ``model_class``.
+    :param mlflow_model: The model configuration to which to add the ``mlflow.pyfunc`` flavor.
+    """
+    return Model.log(artifact_path=artifact_path, flavor=mlflow.pyfunc, artifacts=artifacts,
+                     parameters=parameters, model_class=model_class, conda_env=conda_env,
+                     code_paths=code_paths)
 
 
 def load_pyfunc(path, run_id=None, suppress_warnings=False):
